Don't exclude lines containing only braces...
It's impossible to know when to draw the line on this and it makes
specifying exactly what is considered a SLOC less simple than it should be.
How many braces should be ignored? What about braces followed by semicolons,
as in JavaScript? Just including all non-blank, non-comment lines is much
easier.